Draft: move functions to draftgeoutils.circles, part 1
